Racktables passwords
My first plugin for Racktables on Github.
a simple password tab for managing your passwords.

it is still very basic, for now there is no checkup on special
characters etc. but that will follow in the next version,
also the passwords are stored as  plain text in the database, since MD5
encryption is only one way.

I know its not perfect yet, but there is progress ;-)